requirements.txt
torch
transformers
scikit-learn
pandas
numpy
tiktoken
blobfile
protobuf
kobert-transformers
python-mecab-ko
soynlp